requirements.txt
flask==3.0.0
tensorflow==2.14.0
numpy==1.26.2
pillow==10.0.0
gunicorn==21.2.0
